Resolve corn-flakes.

+++ usr.sbin/amd/amd/nfs_prot_svc.c
@@ -1,7 +1,7 @@
-/*	$NetBSD: nfs_prot_svc.c,v 1.6 1998/08/08 22:33:30 christos Exp $	*/
+/*	$NetBSD: nfs_prot_svc.c,v 1.7 1999/02/01 19:05:10 christos Exp $	*/
 
 /*
- * Copyright (c) 1997-1998 Erez Zadok
+ * Copyright (c) 1997-1999 Erez Zadok
  * Copyright (c) 1989 Jan-Simon Pendry
  * Copyright (c) 1989 Imperial College of Science, Technology & Medicine
  * Copyright (c) 1989 The Regents of the University of California.
@@ -19,7 +19,7 @@
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
  * 3. All advertising materials mentioning features or use of this software
- *    must display the following acknowledgement:
+ *    must display the following acknowledgment:
  *      This product includes software developed by the University of
  *      California, Berkeley and its contributors.
  * 4. Neither the name of the University nor the names of its contributors
@@ -40,7 +40,7 @@
  *
  *      %W% (Berkeley) %G%
  *
- * Id: nfs_prot_svc.c,v 5.2.2.1 1992/02/09 15:09:30 jsp beta 
+ * Id: nfs_prot_svc.c,v 1.3 1999/01/13 23:30:59 ezk Exp 
  *
  */
 
@@ -230,7 +230,7 @@ nfs_program_2(struct svc_req *rqstp, SVCXPRT *transp)
 		   (SVC_IN_ARG_TYPE) &argument)) {
     plog(XLOG_ERROR,
 	 "NFS xdr decode failed for %d %d %d",
-	 rqstp->rq_prog, rqstp->rq_vers, rqstp->rq_proc);
+	 (int) rqstp->rq_prog, (int) rqstp->rq_vers, (int) rqstp->rq_proc);
     svcerr_decode(transp);
     return;
   }
